artifact rejection condition 2
I added the automatic artifact rejection code to the second condition as well and added alternative numbers on the eog code (but it's a mess to try them, will do it again tomorrow).
I changed the data input of ft_rejectartifacts since the new variable gives me an error message!